Teacher tool: set up infrastructure to send messages to iframe (#9811)
* makecodeframe only holds ui, message handling for iframe split into the makecodeeditor service, file made for the transform

* successfully sending message to set high contrast to the iframe when the evaluate button is pressed

* cleaned up a bit

* do something with the results of sending the message

* consolidated ref logic in the service, got rid of extra logic in editor container

* got rid of the editor container, it doesn't do anything meaningful anymore

* got rid of the load event listener

* changed ref handling, some minor changes in the service

* no longer exposing sendMessageAsync

* pushing messages to the queue even if  the iframe isn't ready yet

* changed functions from anonymous to declarative

* fixed set editor ref logic, added check to see if message is undefined

* upgrade codeql analysis to v2
